README.md
This paper discusses the design procedures for both low pass and bandpass filters using microstip line technology. ADS is used for schematic simulation, Momentum is a field solver used for a more accurate characterization, and the filters are both fabricated for measurement with a network analyzer. The results for all cases are presented and discussed.